Put the HeadExt inside Repository to make optimisation possible
